fix(preview): never preview binary files
I think this is a bug fix instead of a new feature.  The reason for this is
that the current behavior _overrides user choice_ of `FZF_PREVIEW_COMMAND`. No
matter what preview command they provide, we never run it on binary files.
However, if the user supplies _no_ preview command, then we'll run it through
`bat` which produces an error.  This sounds like a mistake.

This commit makes the behavior when no preview command is supplied the same as
the behavior when a preview command is given.

Before:
<img width="1251" alt="image" src="https://github.com/user-attachments/assets/d5f66828-5a17-439d-bd62-b670e6e9e2e9" />

After:
<img width="1232" alt="image" src="https://github.com/user-attachments/assets/dd1d811d-fdc0-4023-8bfa-e97e5b9fdca5" />